refactor: 테스트 케이스 수정 및 추가
