+++ src/test/kotlin/blackjack/domain/model/DeckTest.kt
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test
 
 class DeckTest {
     @Test
-    fun `카드들에서 카드를 반환한다`() {
+    fun `덱에서 카드를 반환한다`() {
         val deck = Deck(listOf(Card(Suit.HEART, Rank.SIX)))
         assertThat(deck.draw(1)).isEqualTo(listOf(Card(Suit.HEART, Rank.SIX)))
     }
